requirements.txt
requests>=2.31.0
httpx>=0.25.0
termcolor>=2.3.0
rich>=13.5.2
python-dotenv>=1.0.0
argparse>=1.4.0
python-nmap>=0.7.1 